+++ paper.bib
@@ -12,7 +12,8 @@
     volume=104,
     number=1,
     pages={148--175},
-    year=2016
+    year=2016,
+    doi={10.1109/jproc.2015.2494218}
 }
 
 @book{GarnettR2023book,
@@ -65,7 +66,8 @@
     booktitle={Information Science for Materials Discovery and Design},
     pages={45--75},
     year=2016,
-    publisher={Springer}
+    publisher={Springer},
+    doi={10.1007/978-3-319-23871-5_3}
 }
 
 @article{DurisJ2020phrl,
@@ -75,7 +77,8 @@
     volume=124,
     number=12,
     pages="124801",
-    year=2020
+    year=2020,
+    doi={10.1103/PhysRevLett.124.124801}
 }
 
 @inproceedings{KorovinaK2020aistats,
@@ -99,7 +102,8 @@
     journal={Nature},
     volume=590,
     pages={89--96},
-    year=2021
+    year=2021,
+    doi={10.1038/s41586-021-03213-y}
 }
 
 @article{KimJ2021ml,
@@ -109,7 +113,8 @@
     volume=110,
     number=5,
     pages={857--879},
-    year=2021
+    year=2021,
+    doi={10.1007/s10994-021-05949-0}
 }
 
 @book{RasmussenCE2006book,
@@ -133,7 +138,8 @@
     journal={Nature},
     volume=585,
     pages={357--362},
-    year=2020
+    year=2020,
+    doi={10.1038/s41586-020-2649-2}
 }
 
 @article{VirtanenP2020nm,
@@ -142,7 +148,8 @@
     journal={Nature Methods},
     volume=17,
     pages={261--272},
-    year=2020
+    year=2020,
+    doi={10.1038/s41592-019-0686-2}
 }
 
 @inproceedings{ChoiSCT2022mcqmc,
@@ -168,7 +175,8 @@
     volume=45,
     number=1,
     pages={5--32},
-    year=2001
+    year=2001,
+    doi={10.1007/978-3-030-56485-8_3}
 }
 
 @article{HutterF2014ai,
@@ -177,7 +185,8 @@
     journal={Artificial Intelligence},
     volume=206,
     pages={79--111},
-    year=2014
+    year=2014,
+    doi={10.1016/j.artint.2013.10.003}
 }
 
 @article{MockusJ1978tgo,
@@ -195,7 +204,8 @@
     journal={Journal of Global Optimization},
     volume=13,
     pages={455--492},
-    year=1998
+    year=1998,
+    doi={10.1023/A:1008306431147}
 }
 
 @article{ThompsonWR1933biometrika,
@@ -205,7 +215,8 @@
     volume=25,
     number={3/4},
     pages={285--294},
-    year=1933
+    year=1933,
+    doi={10.2307/2332286}
 }
 
 @article{HuangD2006jgo,
@@ -214,7 +225,8 @@
     journal={Journal of Global Optimization},
     volume=34,
     pages={441--466},
-    year=2006
+    year=2006,
+    doi={10.1007/s10898-005-2454-3}
 }
 
 @inproceedings{SrinivasN2010icml,
@@ -232,7 +244,8 @@
     volume=16,
     number=5,
     pages={1190--1208},
-    year=1995
+    year=1995,
+    doi={10.2172/204262}
 }
 
 @inproceedings{HansenN1997eufit,
@@ -250,5 +263,6 @@
     volume=79,
     number=1,
     pages={157--181},
-    year=1993
+    year=1993,
+    doi={10.1007/bf00941892}
 }
