+++ paper.bib
@@ -13,7 +13,7 @@
     number=1,
     pages={148--175},
     year=2016,
-    doi={10.1109/jproc.2015.2494218}
+    doi={10.1109/JPROC.2015.2494218}
 }
 
 @book{GarnettR2023book,
@@ -245,7 +245,7 @@
     number=5,
     pages={1190--1208},
     year=1995,
-    doi={10.2172/204262}
+    doi={10.1137/0916069}
 }
 
 @inproceedings{HansenN1997eufit,
@@ -264,5 +264,5 @@
     number=1,
     pages={157--181},
     year=1993,
-    doi={10.1007/bf00941892}
+    doi={10.1007/BF00941892}
 }
