+++ paper.bib
@@ -176,7 +176,7 @@
     number=1,
     pages={5--32},
     year=2001,
-    doi={10.1007/978-3-030-56485-8_3}
+    doi={10.1023/A:1010933404324}
 }
 
 @article{HutterF2014ai,
@@ -216,7 +216,7 @@
     number={3/4},
     pages={285--294},
     year=1933,
-    doi={10.2307/2332286}
+    doi={10.1093/biomet/25.3-4.285}
 }
 
 @article{HuangD2006jgo,
