### 2023-01-14 #5045 @arfon
Repository stats for default branch:

```
cloc bayeso/
     194 text files.
     175 unique files.
      32 files ignored.

Language                     files          blank        comment           code
Python                          94           2914           2674          11632
SVG                             22              4              4           9803
Jupyter Notebook                 5              0            517            361
reStructuredText                39            300            806            186
Markdown                         2             44              0            126
YAML                             4             17             27            109
Text                             5              2              0             29
DOS Batch                        1              8              1             27
Bourne Shell                     1              0              0             10
make                             1              4              6             10
INI                              1              0              0              2
SUM:                           175           3293           4035          22295
```
