### 2023-01-14 #5045 @jungtaekkim
@arfon Can I merge a `joss-paper` branch with a working branch (0.5.4)?  My paper describes the working branch and it is not merged to a `main` branch yet.  I am planning to merge it after a review process of JOSS.

Okay, I am looking forward to starting a review process for our paper soon.
