+++ tests/common/test_gp_gp.py
@@ -1,0 +1,211 @@
+#
+# author: Jungtaek Kim (jtkim@postech.ac.kr)
+# last updated: March 22, 2021
+#
+"""test_gp"""
+
+import typing
+import pytest
+import numpy as np
+
+from bayeso import constants
+from bayeso.gp import gp as package_target
+from bayeso.gp import gp_kernel
+from bayeso.utils import utils_covariance
+
+
+TEST_EPSILON = 1e-7
+
+def test_sample_functions_typing():
+    annos = package_target.sample_functions.__annotations__
+
+    assert annos['mu'] == np.ndarray
+    assert annos['Sigma'] == np.ndarray
+    assert annos['num_samples'] == int
+    assert annos['return'] == np.ndarray
+
+def test_sample_functions():
+    num_points = 10
+    mu = np.zeros(num_points)
+    Sigma = np.eye(num_points)
+    num_samples = 5
+
+    with pytest.raises(AssertionError) as error:
+        package_target.sample_functions(mu, 'abc')
+    with pytest.raises(AssertionError) as error:
+        package_target.sample_functions('abc', Sigma)
+    with pytest.raises(AssertionError) as error:
+        package_target.sample_functions(mu, np.eye(20))
+    with pytest.raises(AssertionError) as error:
+        package_target.sample_functions(mu, np.ones(num_points))
+    with pytest.raises(AssertionError) as error:
+        package_target.sample_functions(np.zeros(20), Sigma)
+    with pytest.raises(AssertionError) as error:
+        package_target.sample_functions(np.eye(10), Sigma)
+    with pytest.raises(AssertionError) as error:
+        package_target.sample_functions(mu, Sigma, num_samples='abc')
+    with pytest.raises(AssertionError) as error:
+        package_target.sample_functions(mu, Sigma, num_samples=1.2)
+
+
+    functions = package_target.sample_functions(mu, Sigma, num_samples=num_samples)
+    assert functions.shape[1] == num_points
+    assert functions.shape[0] == num_samples
+
+def test_predict_with_cov_typing():
+    annos = package_target.predict_with_cov.__annotations__
+
+    assert annos['X_train'] == np.ndarray
+    assert annos['Y_train'] == np.ndarray
+    assert annos['X_test'] == np.ndarray
+    assert annos['cov_X_X'] == np.ndarray
+    assert annos['inv_cov_X_X'] == np.ndarray
+    assert annos['hyps'] == dict
+    assert annos['str_cov'] == str
+    assert annos['prior_mu'] == typing.Union[typing.Callable, type(None)]
+    assert annos['debug'] == bool
+    assert annos['return'] == typing.Tuple[np.ndarray, np.ndarray, np.ndarray]
+
+def test_predict_with_cov():
+    np.random.seed(42)
+    dim_X = 2
+    num_X = 5
+    num_X_test = 20
+    X = np.random.randn(num_X, dim_X)
+    Y = np.random.randn(num_X, 1)
+    X_test = np.random.randn(num_X_test, dim_X)
+    prior_mu = None
+    cov_X_X, inv_cov_X_X, hyps = gp_kernel.get_optimized_kernel(X, Y, prior_mu, 'se')
+
+    with pytest.raises(AssertionError) as error:
+        package_target.predict_with_cov(X, Y, X_test, cov_X_X, inv_cov_X_X, hyps, str_cov='se', prior_mu='abc')
+    with pytest.raises(AssertionError) as error:
[…]
