### 2023-02-21 a732960
Reduce randomness due to scipy.minimize
