+++ bayeso/constants.py
@@ -16,6 +16,8 @@ JITTER_LOG = 1e-7
 
 TOLERANCE_DUPLICATED_ACQ = 1e-8
 
+FLOAT_CLIPPING = '{:.6f}'
+
 STR_SURROGATE = 'gp'
 STR_SURROGATE_TREES = 'rf'
 STR_OPTIMIZER_METHOD_GP = 'BFGS'
@@ -139,3 +141,4 @@ TYPING_UNION_ARRAY_NONE = typing.Union[TYPE_ARR, TYPE_NONE]
 TYPING_UNION_ARRAY_FLOAT = typing.Union[TYPE_ARR, float]
 TYPING_UNION_CALLABLE_NONE = typing.Union[TYPING_CALLABLE, TYPE_NONE]
 TYPING_UNION_STR_NONE = typing.Union[str, TYPE_NONE]
+TYPING_UNION_FLOAT_ARRAY = typing.Union[str, TYPE_ARR]
