### 2023-02-22 0cd2ab9
Remove fun_objective for get_samples('grid', )
