### 2023-02-22 1596c3f
Remove randomness by adding a seed for optimize()
