+++ bayeso/constants.py
@@ -16,8 +16,6 @@ JITTER_LOG = 1e-7
 
 TOLERANCE_DUPLICATED_ACQ = 1e-8
 
-FLOAT_CLIPPING = '{:.6f}'
-
 STR_SURROGATE = 'gp'
 STR_SURROGATE_TREES = 'rf'
 STR_OPTIMIZER_METHOD_GP = 'BFGS'

[…]
