### 2023-02-23 ec1d470
Update bo_w_tp to pass seed
