### 2023-03-31 #5045 @drvinceknight
@thomaspinder (sorry for the "re"ping but this does seem right in your area of expertise) @jmp75 @BenWinchester I am pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “BayesO: A Bayesian optimization framework in Python”. You can find more information at the top of this Github issue.
