### 2023-03-31 #5320 @editorialbot
[REVIEW]: BayesO: A Bayesian optimization framework in Python
**Submitting author:** @jungtaekkim (<a href="http://orcid.org/0000-0002-1905-1399">Jungtaek Kim</a>)
**Repository:** https://github.com/jungtaekkim/bayeso
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.5.4
