+++ tests/common/test_bo_bo_w_tp.py
@@ -154,19 +154,19 @@ def test_get_samples():
 
     truth_arr_initials = np.array([
         [
-            8.78516613971442,
-            -0.6113853892311454,
-            -4.274874331895262,
+            4.31029474362731,
+            1.257471889257431,
+            3.06412766687572,
         ],
         [
-            2.7565963030792773,
-            1.627942705526948,
-            1.6141902864910662,
+            8.698135614395142,
+            -0.250022292137146,
+            -0.012653172016143799,
         ],
         [
-            1.0978685109876096,
-            -1.511254413984716,
-            -1.5049133892171085,
+            5.779154300689697,
+            0.04064440727233887,
+            2.2647011280059814,
         ],
     ])
 
@@ -182,19 +182,19 @@ def test_get_samples():
 
     truth_arr_initials = np.array([
         [
-            3.3124358790165855,
-            -1.4099903230606423,
-            -0.4462920191434243,
+            9.486941305084901,
+            1.1840371390061812,
+            4.866438875059044,
         ],
         [
-            8.312435879016585,
-            1.2566763436060246,
-            -2.446292019143424,
+            0.4244413050849005,
+            -0.4455924906234483,
+            0.8664388750590444,
         ],
         [
-            0.8124358790165853,
-            -0.07665698972730861,
-            -4.446292019143424,
+            5.4244413050849,
+            -1.7789258239567811,
+            2.8664388750590453,
         ],
     ])
 
@@ -245,25 +245,26 @@ def test_get_initials():
     arr_initials = model_bo.get_initials('sobol', 3)
     arr_initials = model_bo.get_initials('sobol', 3, seed=42)
 
+    print('sobol')
     for elem_1 in arr_initials:
         for elem_2 in elem_1:
             print(elem_2)
 
     truth_arr_initials = np.array([
         [
-            8.78516613971442,
-            -0.6113853892311454,
-            -4.274874331895262,
+            4.31029474362731,
+            1.257471889257431,
+            3.06412766687572,
         ],
         [
-            2.7565963030792773,
-            1.627942705526948,
-            1.6141902864910662,
+            8.698135614395142,
+            -0.250022292137146,
+            -0.012653172016143799,
         ],
         [
-            1.0978685109876096,
-            -1.511254413984716,
-            -1.5049133892171085,
+            5.779154300689697,
+            0.04064440727233887,
+            2.2647011280059814,
         ],
     ])
 
@@ -786,54 +787,54 @@ def test_compute_acquisitions():
 
     truth_X_test = np.array([
         [
-            3.328958908095956,
-            -1.8729291455820203,
-            0.2839687094092369,
+            3.359774835407734,
+            -0.7351906783878803,
+            -3.654018910601735,
         ],
         [
-            8.11741182114929,
-            0.3799784183502197,
-            -0.05574141861870885,
+            5.692976117134094,
+            0.06583881378173828,
+            4.514206647872925,
         ],
         [
-            6.735238193068653,
-            -0.9264274807646871,
-            3.631770429201424,
+            9.900951385498047,
+            -1.9652910344302654,
+            -2.2755324840545654,
         ],
         [
-            2.13300823001191,
-            1.3245289996266365,
-            -3.547573888208717,
+            2.2963321208953857,
+            1.359175443649292,
+            0.7950365543365479,
         ],
         [
-            0.6936023756861687,
-            -0.018464308232069016,
-            -2.1043178741820157,
+            0.2561802417039871,
+            -1.420634150505066,
+            3.1929773092269897,
         ],
         [
-            5.438151848502457,
-            1.7285785367712379,
-            2.0298107899725437,
+            7.546542286872864,
[…]

[…]
