### 2023-06-02 d1d0921
Remove python 3.6 support and update github actions
