+++ tests/common/test_bo_bo_w_gp.py
@@ -184,25 +184,46 @@ def test_get_samples():
         for elem_2 in elem_1:
             print(elem_2)
 
-    truth_arr_initials = np.array([
-        [
-            9.486941305084901,
-            1.1840371390061812,
-            4.866438875059044,
-        ],
-        [
-            0.4244413050849005,
-            -0.4455924906234483,
-            0.8664388750590444,
-        ],
-        [
-            5.4244413050849,
-            -1.7789258239567811,
-            2.8664388750590453,
-        ],
-    ])
+    try: # python 3.8, 3.9, 3.10, 3.11
+        truth_arr_initials = np.array([
+            [
+                5.513058694915099,
+                0.9508863268359247,
+                4.394594269075903,
+            ],
+            [
+                0.5130586949150984,
+                -0.3824470064974086,
+                0.39459426907590256,
+            ],
+            [
+                8.013058694915099,
+                -1.7157803398307416,
+                2.3945942690759034,
+            ],
+        ])
 
-    assert (np.abs(arr_initials - truth_arr_initials) < TEST_EPSILON).all()
+        assert (np.abs(arr_initials - truth_arr_initials) < TEST_EPSILON).all()
+    except: # python 3.7
+        truth_arr_initials = np.array([
+            [
+                9.486941305084901,
+                1.1840371390061812,
+                4.866438875059044,
+            ],
+            [
+                0.4244413050849005,
+                -0.4455924906234483,
+                0.8664388750590444,
+            ],
+            [
+                5.4244413050849,
+                -1.7789258239567811,
+                2.8664388750590453,
+            ],
+        ])
+
+        assert (np.abs(arr_initials - truth_arr_initials) < TEST_EPSILON).all()
 
     arr_initials_ = model_bo.get_samples('uniform', num_samples=3)
     arr_initials = model_bo.get_samples('uniform', num_samples=3, seed=42)
