+++ bayeso/constants.py
@@ -139,3 +139,4 @@ TYPING_UNION_ARRAY_NONE = typing.Union[TYPE_ARR, TYPE_NONE]
 TYPING_UNION_ARRAY_FLOAT = typing.Union[TYPE_ARR, float]
 TYPING_UNION_CALLABLE_NONE = typing.Union[TYPING_CALLABLE, TYPE_NONE]
 TYPING_UNION_STR_NONE = typing.Union[str, TYPE_NONE]
+TYPING_UNION_FLOAT_ARRAY = typing.Union[str, TYPE_ARR]
