+++ paper.bib
@@ -152,14 +152,6 @@
     doi={10.1038/s41592-019-0686-2}
 }
 
-@inproceedings{ChoiSCT2022mcqmc,
-    author={{Choi}, S.-C. T. and {Hickernell}, F. J. and {Jagadeeswaran}, R. and {McCourt}, M. J. and {Sorokin}, A. G.},
-    title={Quasi-{Monte Carlo} Software},
-    booktitle={Proceedings of the International Conference on Monte Carlo and Quasi-Monte Carlo Methods in Scientific Computing (MCQMC)},
-    pages={23--47},
-    year=2022
-}
-
 @misc{HansenN2019software,
     author={{Hansen}, N. and {Akimoto}, Y. and {Baudis}, P.},
     title={{CMA-ES/pycma} on {GitHub}},
@@ -265,4 +257,14 @@
     pages={157--181},
     year=1993,
     doi={10.1007/BF00941892}
+}
+
+@article{RoyPT2023joss,
+    author={{Roy}, P. T. and {Owen}, A. B. and {Balandat}, M. and {Haberland}, M.},
+    title={Quasi-{Monte Carlo} Methods in {Python}},
+    journal={Journal of Open Source Software},
+    volume=8,
+    number=84,
+    pages=5309,
+    year=2023
 }
