### 2023-06-03 #5320 @jungtaekkim
Dear editor and reviewers,

we recently faced an installation issue of `qmcpy`.  To resolve such an issue and make dependency much simpler, we replaced a quasi-Monte Carlo package, which was `qmcpy`, with [a submodule in SciPy](https://joss.theoj.org/papers/10.21105/joss.05309.pdf), which is recently included in SciPy.

We have reflected this in the `joss-paer` branch.  Please consider it in a reviewing process.

Best regards,
Authors of a paper "BayesO: A Bayesian optimization framework in Python".
