+++ paper.bib
@@ -266,5 +266,6 @@
     volume=8,
     number=84,
     pages=5309,
-    year=2023
+    year=2023,
+    doi={10.21105/joss.05309}
 }
