+++ _config.yml
@@ -10,7 +10,7 @@ relative_permalinks: true
 title:            Hyde
 tagline:          'A Jekyll theme'
 description:      'A brazen two-column <a href="http://jekyllrb.com" target="_blank">Jekyll</a> theme that pairs a prominent sidebar with uncomplicated content. Made by <a href="https://twitter.com/mdo" target="_blank">@mdo</a>.'
-url:              https://bayeso.org
+url:              http://hyde.getpoole.com
 baseurl:          /
 
 author:
