+++ _config.yml
@@ -4,23 +4,22 @@ highlighter:      pygments
 
 # Permalinks
 permalink:        pretty
-relative_permalinks: true
 
 # Setup
-title:            Hyde
+title:            BayesO
 tagline:          'A Jekyll theme'
 description:      'A brazen two-column <a href="http://jekyllrb.com" target="_blank">Jekyll</a> theme that pairs a prominent sidebar with uncomplicated content. Made by <a href="https://twitter.com/mdo" target="_blank">@mdo</a>.'
-url:              http://hyde.getpoole.com
+url:              https://bayeso.org
 baseurl:          /
+paginate:         5
 
 author:
-  name:           'Mark Otto'
+  name:           'Jungtaek Kim'
   url:            https://twitter.com/mdo
 
-paginate:         5
 
 # Custom vars
 version:          2.1.0
 
 github:
-  repo:           https://github.com/poole/hyde
+  repo:           https://github.com/jungtaekkim/bayeso
