+++ _config.yml
@@ -9,7 +9,7 @@ permalink:        pretty
 title:            BayesO
 description:      'Simple, but essential Bayesian optimization package'
 url:              https://bayeso.org
-baseurl:          https://bayeso.org
+baseurl:          /
 paginate:         5
 
 author:
