+++ README.md
@@ -1,0 +1,8 @@
+# GitHub Pages for BayesO
+
+This page is built with [Hyde](https://github.com/poole/hyde), which is licensed under the [MIT License](LICENSE.md).
+
+Hyde is a brazen two-column [Jekyll](http://jekyllrb.com) theme that pairs a prominent sidebar with uncomplicated content.
+It is based on [Poole](http://getpoole.com), the Jekyll butler.
+
+The author of Hyde is Mark Otto, who is available at [GitHub](https://github.com/mdo) or [Twitter](https://twitter.com/mdo).
