+++ _config.yml
@@ -16,7 +16,6 @@ author:
   name:           'Jungtaek Kim'
   url:            https://jungtaek.github.io
 
-
 # Custom vars
 version:          0.5.3
 
