### 2023-07-03 #5320 @jungtaekkim
Dear editor @drvinceknight and reviewers @thomaspinder @salrm8,

First of all, we really appreciate your service for JOSS.

We think that it has been six months since we submitted our paper.  We therefore would like to ask when we will be able to get the decision on the paper.

Thank you again and we hope to get the decision soon.

Best regards,
Authors of a paper "BayesO: A Bayesian optimization framework in Python".
