### 2023-08-16 #5320 @thomaspinder
@drvinceknight I have opened issues in Bayeso now with specific comments and a high-level review. Once these issues have been resolved, my checklist will be complete.

To the authors, I would like to apologise for my delay in getting this review done.
