+++ bayeso/constants.py
@@ -31,7 +31,7 @@ STR_MLM_METHOD = 'regular'
 STR_MODELSELECTION_METHOD = 'ml'
 
 NUM_GRIDS_AO = 50
-NUM_SAMPLES_AO = 100
+NUM_SAMPLES_AO = 128
 NUM_SAMPLES_AO_TREES = 5000
 
 MULTIPLIER_ACQ = 10.0
