### 2023-08-16 33c487c
Add ISSUE_TEMPLATE and update CODE_OF_CONDUCT
