### 2023-08-16 5aba47a
Clarify a sentence for code analysis
