### 2023-08-16 fb78086
Clarify a sentence in installation
