### 2023-08-16 b222010
Fix tests for halton sequences
