### 2023-08-16 91321e2
Fix tests for halton sequences
