+++ tests/common/test_bo_bo_w_trees.py
@@ -205,19 +205,19 @@ def test_get_samples():
     else:
         truth_arr_initials = np.array([
             [
-                5.513058694915099
-                -1.3929280802587178
-                -3.572948073154651
+                5.513058694915099,
+                -1.3929280802587178,
+                -3.572948073154651,
             ],
             [
-                0.5130586949150984
-                1.2737385864079487
-                0.4270519268453521
+                0.5130586949150984,
+                1.2737385864079487,
+                0.4270519268453521,
             ],
             [
-                8.013058694915099
-                -0.059594746925384356
-                2.427051926845353
+                8.013058694915099,
+                -0.059594746925384356,
+                2.427051926845353,
             ],
         ])
 
