### 2023-08-16 ff7d290
Fix tests for halton sequences
