### 2023-08-16 b5c270b
Fix tests for halton sequences
