+++ paper.bib
@@ -269,3 +269,11 @@
     year=2023,
     doi={10.21105/joss.05309}
 }
+
+@misc{tqdm2016software,
+    author={The tqdm authors},
+    title={tqdm},
+    publisher={GitHub},
+    url={https://github.com/tqdm/tqdm},
+    year=2016
+}
