+++ paper.bib
@@ -271,8 +271,8 @@
 }
 
 @misc{tqdm2016software,
-    author={The tqdm authors},
-    title={tqdm},
+    author={{The tqdm authors}},
+    title={{tqdm}},
     publisher={GitHub},
     url={https://github.com/tqdm/tqdm},
     year=2016
