+++ pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-timeout = 25
+timeout = 30
