+++ CONTRIBUTING.md
@@ -1,0 +1,31 @@
+We deeply appreciate your interest in the BayesO project :smile:.
+
+This document describes how you can contribute to the project.
+
+## Code of Conduct
+
+First of all, we follow the [code of conduct](CODE_OF_CONDUCT.md) in order to create a positive, diverse, and inclusive open-source community.
+
+Please read the linked document carefully and let us know if you see any violation against the code of conduct.
+
+## How Can I Contribute?
+
+We welcome any form of contributions.
+Here some possible formats are introduced.
+
+### Issues
+
+You can post issues on bug reports, feature requests, and others via [this link](https://github.com/jungtaekkim/bayeso/issues).
+
+We provided issue templates but you can start from a blank template.
+
+### Pull Requests
+
+You can open pull requests via [this link](https://github.com/jungtaekkim/bayeso/pulls).
+
+Any pull requests on fixing bugs and suggesting new features are welcomed.
+To create a decent pull request, you need to consider consistency on code conventions and completeness of implementation.
+We would like to let you know that we will review your pull request.
+Also, we will be able to ask you to modify the pull request or modify it ourselves.
+
+Thank you again!
