+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -8,13 +8,16 @@ assignees: ''
 ---
 
 **Is your feature request related to a problem? Please describe.**
-A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+A clear and concise description of what the problem is.
 
-**Describe the solution you'd like**
+**Describe the solution you would like**
 A clear and concise description of what you want to happen.
 
-**Describe alternatives you've considered**
-A clear and concise description of any alternative solutions or features you've considered.
+**Describe alternatives you have considered**
+A clear and concise description of any alternative solutions or features you have considered.
+
+**Provide the references we can refer to**
+References (e.g., publications and technical reports) that present the methodology you would like.
 
 **Additional context**
 Add any other context or screenshots about the feature request here.
