+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -13,9 +13,8 @@ A clear and concise description of what the bug is.
 **To Reproduce**
 Steps to reproduce the behavior:
 1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error
+2. Command '....'
+3. See error
 
 **Expected behavior**
 A clear and concise description of what you expected to happen.
@@ -23,10 +22,10 @@ A clear and concise description of what you expected to happen.
 **Screenshots**
 If applicable, add screenshots to help explain your problem.
 
-**Desktop (please complete the following information):**
- - OS: [e.g. Ubuntu 23.04]
- - Python Version: [e.g. 3.11.4]
- - BayesO Version [e.g. 0.5.4]
+**Machine:**
+ - OS: [e.g., Ubuntu 23.04]
+ - Python Version: [e.g., 3.11.4]
+ - BayesO Version [e.g., 0.5.4]
 
 **Additional context**
 Add any other context about the problem here.
