### 2023-08-16 1c76a64
Update examples; not always save figures
