+++ examples/01_basics/example_basics_gp.py
@@ -1,6 +1,7 @@
-# example_basics_gp
+#
 # author: Jungtaek Kim (jtkim@postech.ac.kr)
-# last updated: April 28, 2020
+# last updated: August 16, 2023
+#
 
 import numpy as np
 import os
@@ -10,9 +11,7 @@ from bayeso.utils import utils_common
 from bayeso.utils import utils_plotting
 
 
-PATH_SAVE = '../figures/gp/'
-
-def main():
+def main(path_save):
     X_train = np.array([
         [-3],
         [-1],
@@ -30,10 +29,12 @@ def main():
         'noise': 0.02,
     }
     mu, sigma, Sigma = gp.predict_with_hyps(X_train, Y_train, X_test, hyps)
-    utils_plotting.plot_gp_via_distribution(X_train, Y_train, X_test, mu, sigma, Y_test, path_save=PATH_SAVE, str_postfix='cos')
+    utils_plotting.plot_gp_via_distribution(X_train, Y_train, X_test, mu, sigma, Y_test, path_save=path_save, str_postfix='cos')
 
 
 if __name__ == '__main__':
-    if not os.path.isdir(PATH_SAVE):
-        os.makedirs(PATH_SAVE)
-    main()
+    path_save = None
+
+    if path_save is not None and not os.path.isdir(path_save):
+        os.makedirs(path_save)
+    main(path_save)
