### 2023-08-16 30d9aa9
Update examples; not always save figures
