+++ examples/02_surrogates/example_generic_trees.py
@@ -1,6 +1,7 @@
-# example_generic_trees
+#
 # author: Jungtaek Kim (jtkim@postech.ac.kr)
-# last updated: August 6, 2021
+# last updated: August 16, 2023
+#
 
 import numpy as np
 import os
@@ -10,10 +11,8 @@ from bayeso.trees import trees_common
 from bayeso.utils import utils_common
 from bayeso.utils import utils_plotting
 
-PATH_SAVE = '../figures/rf/'
 
-
-def main():
+def main(path_save):
     np.random.seed(42)
     X_train = np.array([
         [-3.0],
@@ -43,11 +42,12 @@ def main():
 
     mu, sigma = trees_common.predict_by_trees(X_test, trees)
 
-    utils_plotting.plot_gp_via_distribution(X_train, Y_train, X_test, mu, sigma, Y_test, path_save=PATH_SAVE, str_postfix='cos')
+    utils_plotting.plot_gp_via_distribution(X_train, Y_train, X_test, mu, sigma, Y_test, path_save=path_save, str_postfix='cos')
 
 
 if __name__ == '__main__':
-    if not os.path.isdir(PATH_SAVE):
-        os.makedirs(PATH_SAVE)
+    path_save = None
 
-    main()
+    if path_save is not None and not os.path.isdir(path_save):
+        os.makedirs(path_save)
+    main(path_save)
