+++ paper.bib
@@ -277,3 +277,62 @@
     url={https://github.com/tqdm/tqdm},
     year=2016
 }
+
+@misc{gpy2012software,
+    author={{GPy}},
+    title={{GPy}: A {Gaussian} process framework in {Python}},
+    publisher={GitHub},
+    url={https://github.com/SheffieldML/GPy},
+    year=2012
+}
+
+@misc{gpyopt2016software,
+    author={{The GPyOpt authors}},
+    title={{GPyOpt}: A {Bayesian} Optimization Framework in {Python}},
+    publisher={GitHub},
+    url={https://github.com/SheffieldML/GPyOpt},
+    year=2016
+}
+
+@article{LindauerM2022jmlr,
+    author={Lindauer, M. and Eggensperger, K. and Feurer, M. and Biedenkapp, A. and Deng, D. and Benjamins, C. and Ruhkopf, T. and Sass, R. and Hutter, F.},
+    title={{SMAC3}: A versatile {Bayesian} optimization package for hyperparameter optimization},
+    journal={Journal of Machine Learning Research},
+    volume=23,
+    number=54,
+    pages={1--9},
+    year=2022
+}
+
+@inproceedings{PaszkeA2019neurips,
+    author={Paszke, A. and Gross, S. and Massa, F. and Lerer, A. and Bradbury, J. and Chanan, G. and Killeen, T. and Lin, Z. and Gimelshein, N. and Antiga, L. and Desmaison, A. and {K\"{o}pf}, A. and Yang, E. and DeVito, Z. and Raison, M. and Tejani, A. and Chilamkurthy, S. and Steiner, B. and Fang, L. and Bai, J. and Chintala, S.},
+    title={{PyTorch}: An imperative style, high-performance deep learning library},
+    booktitle={Advances in Neural Information Processing Systems (NeurIPS)},
+    volume=32,
+    pages={8026--8037},
+    year=2019
+}
+
+@inproceedings{AbadiM2016osdi,
+    author={Abadi, M. and Barham, P. and Chen, J. and Chen, Z. and Davis, A. and Dean, J. and Devin, M. and Ghemawat, S. and Irving, G. and Isard, M. and Kudlur, M. and Levenberg, J. and Monga, R. and Moore, S. and Murray, D. G. and Steiner, B. and Tucker, P. and Vasudevan, V. and Warden, P. and Wicke, M. and Yu, Y. and Zheng, X.},
+    title={{TensorFlow}: A system for large-scale machine learning},
+    booktitle={USENIX Symposium on Operating Systems Design and Implementation (OSDI)},
+    pages={265--283},
+    year=2016
+}
+
+@article{KnuddeN2017arxiv,
+    author={Knudde, N. and {van der Herten}, J. and Dhaene, T. and Couckuyt, I.},
+    title={{GPflowOpt}: A {Bayesian} Optimization Library using {TensorFlow}},
+    journal={{arXiv preprint arXiv}:1711.03845},
+    year=2017
+}
+
+@inproceedings{BalandatM2020neurips,
+    author={Balandat, M. and Karrer, B. and Jiang, D. R. and Daulton, S. and Letham, B. and Wilson, A. G. and Bakshy, E.},
+    title={{BoTorch}: A Framework for Efficient {Monte-Carlo} {Bayesian} Optimization},
+    booktitle={Advances in Neural Information Processing Systems (NeurIPS)},
+    volume=33,
+    pages={21524--21538},
+    year=2020,
+}
