### 2023-08-17 6542d99
Polish paper.md
