+++ paper.md
@@ -103,8 +103,7 @@ that, for this example, an objective function is
 \end{equation}
 where $x \in [-5, 5]$, and Gaussian process regression and expected
 improvement are used as a surrogate model and an acquisition function,
-respectively. To focus on the BayesO system, we omit the details of
-surrogate models and acquisition functions here.
+respectively.
 
 Several related projects on Bayesian optimization have been proposed.
 GPyOpt [@gpyopt2016software], built on GPy [@gpy2012software], has been
@@ -116,12 +115,12 @@ BoTorch [@BalandatM2020neurips] has been introduced.
 
 # Overview of BayesO
 
+![Logo of BayesO\label{fig:logo_bayeso}](figures/logo_bayeso_capitalized.png){ width=50% }
+
 In this section we cover the overview of BayesO including probabilistic
 regression models and acquisition functions, by referring to BayesO v0.5.4.
 For higher versions of BayesO, see official documentation. Note that the
 BayesO logo is shown in \autoref{fig:logo_bayeso}.
-
-![Logo of BayesO.\label{fig:logo_bayeso}](figures/logo_bayeso_capitalized.png){ width=50% }
 
 BayesO supports the following probabilistic regression models:
 
@@ -148,7 +147,7 @@ and L-BFGS-B [@ByrdRH1995siamjsc]. In addition, we also include Thompson
 sampling [@ThompsonWR1933biometrika] in BayesO.
 
 Furthermore, to support an easy-to-use interface, we implement the wrappers
-of Bayesian optimization for particular scenarios as follows:
+of Bayesian optimization for particular scenarios:
 
 - a run with randomly-chosen initial inputs;
 - a run with initial inputs provided;
@@ -163,7 +162,7 @@ management packages.
 inspected to satisfy the code conventions predefined in our software.
 - Type hints: As supported in Python 3, we provide type hints for any
 arguments.
-- Unit tests: All the unit tests for our software are included. We have
+- Unit tests: Unit tests for our software are included. We have
 achieved 100\% coverage. In addition, unit tests for measuring execution
 time are also provided.
 - Dependency: Our package depends on NumPy [@HarrisCR2020nature],
