### 2023-08-17 efd641d
Update examples; not always save figures
