+++ examples/06_hpo/example_hpo_xgboost_ei.py
@@ -1,6 +1,7 @@
-# example_hpo_xgboost_ei
+#
 # author: Jungtaek Kim (jtkim@postech.ac.kr)
-# last updated: December 29, 2020
+# last updated: August 17, 2023
+#
 
 import numpy as np
 import os
@@ -22,15 +23,14 @@ DIGITS_DATA = DIGITS.images
 DIGITS_DATA = np.reshape(DIGITS_DATA, (DIGITS_DATA.shape[0], DIGITS_DATA.shape[1] * DIGITS_DATA.shape[2]))
 DIGITS_LABELS = DIGITS.target
 DATA_TRAIN, DATA_TEST, LABELS_TRAIN, LABELS_TEST = sklearn.model_selection.train_test_split(DIGITS_DATA, DIGITS_LABELS, test_size=0.3, stratify=DIGITS_LABELS)
-PATH_SAVE = '../figures/hpo/'
 
 def fun_target(X):
     print(X)
-    xgb_model = xgb.XGBClassifier(max_depth=int(X[0]), n_estimators=int(X[1]), use_label_encoder=False).fit(DATA_TRAIN, LABELS_TRAIN, eval_metric='mlogloss')
+    xgb_model = xgb.XGBClassifier(max_depth=int(X[0]), n_estimators=int(X[1]), eval_metric='mlogloss').fit(DATA_TRAIN, LABELS_TRAIN)
     preds = xgb_model.predict(DATA_TEST)
     return 1.0 - sklearn.metrics.accuracy_score(LABELS_TEST, preds)
 
-def main():
+def main(path_save):
     # (max_depth, n_estimators)
     num_init = 1
 
@@ -46,11 +46,13 @@ def main():
     arr_time = np.array(list_time)
     arr_time = np.expand_dims(arr_time, axis=0)
 
-    utils_plotting.plot_minimum_vs_iter(arr_Y, ['xgboost'], num_init, True, path_save=PATH_SAVE, str_postfix='xgboost')
-    utils_plotting.plot_minimum_vs_time(arr_time, arr_Y, ['xgboost'], num_init, True, path_save=PATH_SAVE, str_postfix='xgboost')
+    utils_plotting.plot_minimum_vs_iter(arr_Y, ['xgboost'], num_init, True, path_save=path_save, str_postfix='xgboost')
+    utils_plotting.plot_minimum_vs_time(arr_time, arr_Y, ['xgboost'], num_init, True, path_save=path_save, str_postfix='xgboost')
 
 
 if __name__ == '__main__':
-    if not os.path.isdir(PATH_SAVE):
-        os.makedirs(PATH_SAVE)
-    main()
+    path_save = None
+
+    if path_save is not None and not os.path.isdir(path_save):
+        os.makedirs(path_save)
+    main(path_save)
