### 2023-08-17 3b67844
Update examples; not always save figures
