+++ examples/05_benchmarks/example_benchmarks_bohachevsky_bo_ei.py
@@ -1,6 +1,7 @@
-# example_benchmarks_bohachevsky_bo_ei
+#
 # author: Jungtaek Kim (jtkim@postech.ac.kr)
-# last updated: July 8, 2021
+# last updated: August 17, 2023
+#
 
 import numpy as np
 import os
@@ -13,7 +14,6 @@ from bayeso.utils import utils_plotting
 
 
 STR_FUN_TARGET = 'bohachevksy'
-PATH_SAVE = '../figures/benchmarks/'
 
 obj_fun = Bohachevsky()
 
@@ -21,7 +21,7 @@ obj_fun = Bohachevsky()
 def fun_target(X):
     return obj_fun.output(X)
 
-def main():
+def main(path_save):
     num_bo = 5
     num_iter = 10
     num_init = 5
@@ -59,11 +59,13 @@ def main():
     times = np.array(list_time)
     times = np.expand_dims(times, axis=0)
 
-    utils_plotting.plot_minimum_vs_iter(Ys, [STR_FUN_TARGET], num_init, True, path_save=PATH_SAVE, str_postfix=STR_FUN_TARGET)
-    utils_plotting.plot_minimum_vs_time(times, Ys, [STR_FUN_TARGET], num_init, True, path_save=PATH_SAVE, str_postfix=STR_FUN_TARGET)
+    utils_plotting.plot_minimum_vs_iter(Ys, [STR_FUN_TARGET], num_init, True, path_save=path_save, str_postfix=STR_FUN_TARGET)
+    utils_plotting.plot_minimum_vs_time(times, Ys, [STR_FUN_TARGET], num_init, True, path_save=path_save, str_postfix=STR_FUN_TARGET)
 
 
 if __name__ == '__main__':
-    if not os.path.isdir(PATH_SAVE):
-        os.makedirs(PATH_SAVE)
-    main()
+    path_save = None
+
+    if path_save is not None and not os.path.isdir(path_save):
+        os.makedirs(path_save)
+    main(path_save)
