+++ examples/05_benchmarks/example_benchmarks_branin_ts.py
@@ -1,6 +1,7 @@
-# example_benchmarks_branin_ts
+#
 # author: Jungtaek Kim (jtkim@postech.ac.kr)
-# last updated: December 19, 2021
+# last updated: August 17, 2023
+#
 
 import numpy as np
 import os
@@ -12,7 +13,6 @@ from bayeso.utils import utils_plotting
 
 
 STR_FUN_TARGET = 'branin'
-PATH_SAVE = '../figures/benchmarks/'
 
 obj_fun = Branin()
 
@@ -20,7 +20,7 @@ obj_fun = Branin()
 def fun_target(X):
     return obj_fun.output(X)
 
-def main():
+def main(path_save):
     num_bo = 5
     num_init = 1
     num_iter = 50
@@ -44,10 +44,12 @@ def main():
 
     arr_Y = np.array(list_Y)
     arr_Y = np.expand_dims(np.squeeze(arr_Y), axis=0)
-    utils_plotting.plot_minimum_vs_iter(arr_Y, [STR_FUN_TARGET], num_init, True, path_save=PATH_SAVE, str_postfix=STR_FUN_TARGET)
+    utils_plotting.plot_minimum_vs_iter(arr_Y, [STR_FUN_TARGET], num_init, True, path_save=path_save, str_postfix=STR_FUN_TARGET)
 
 
 if __name__ == '__main__':
-    if not os.path.isdir(PATH_SAVE):
-        os.makedirs(PATH_SAVE)
-    main()
+    path_save = None
+
+    if path_save is not None and not os.path.isdir(path_save):
+        os.makedirs(path_save)
+    main(path_save)
