### 2023-08-17 #5320 @salrm8
I apologize for the delay in sending my comments. The work is really nice, congrats to the authors.

Comments:

**Paper**:

    • Line 16: Optimizer -> optimum
    • Line 16: “...we cannot employ any knowledge in … “. This sounds a bit vague to me. Could you please clarify about “knowledge”?
    • A drawback of the paper is that there is no reference to the existing libraries for Bayesian optimization. The authors are encouraged to address this, and highlight the extra features provided by bayesO.
    • Line 40: Perhaps good to revise as “probabilistic regression model (surrogate)”.
    • Line 44: Please revise as “Evaluating the model at a query point”
    • Line 87: What is "tqdm"? Please also add a reference for it.
    • Line 92: Better to replace “our own” by “a”.
    • Figure 1: Assuming the iterations did not stop at 9, please mention this in the caption for clarity.

**General:**
Are the BO methods implemented in the software able to handle uncertain data?

**Code:**

    • Please add a list of tests.
    • Add a list of examples
    • Please make the tests automatic
    • In `/exmaples/02_surrogates/`, there is an error when running `example_gp_mml_comparisons.py`:
      `ModuleNotFoundError: No module named 'bayeso_benchmarks'`
    • The same error in /examples/07_wandb/, when running wandb_branin.py
    • The same error for notebook `example_bo_branin.ipynb`
    • Is `xgboost` among the requirements? It is not installed automatically although it is needed for running `example_hpo_xgboost_ei.py`.
    • Could the authors provide an example (surrogate and BO) for a 2D parameter space? It seems, the examples are only based on one parameter.
    • Adding a bit of text in the jupyter notebooks in `/examples/99_notebooks/` would be helpful to understand the topic of each notebook.
    • The documentation is good, but can be improved by adding a list for the examples and considering the previous point. Another improvement would be adding a bit of theory.
