### 2023-08-20 40cc96c
Update gp.rst
