+++ docs/example/gp.rst
@@ -2,7 +2,12 @@ Building Gaussian Process Regression
 ====================================
 
 This example is for building Gaussian process regression models.
-First of all, import the packages we need and **bayeso**.
+In this example, we cover three scenarios:
+Gaussian process with fixed hyperparameters,
+Gaussian process with learned hyperparameters,
+and Gaussian process with particular priors.
+
+First of all, import the package we need and **bayeso**.
 
 .. code-block:: python
 
@@ -14,6 +19,9 @@ First of all, import the packages we need and **bayeso**.
     from bayeso.utils import utils_plotting
 
 Declare some parameters to control this example.
+`use_tex` is a plotting style for LaTeX,
+`num_test` is the number of test data points,
+and `str_cov` is a kernel choice.
 
 .. code-block:: python
 
@@ -21,7 +29,8 @@ Declare some parameters to control this example.
     num_test = 200
     str_cov = 'matern52'
 
-Make a simple synthetic dataset, which produces with cosine functions.
+Make a simple synthetic dataset, which is produced with a cosine function.
+The underlying true function is $\cos(x) + 10$.
 
 .. code-block:: python
 
@@ -39,6 +48,7 @@ Make a simple synthetic dataset, which produces with cosine functions.
     Y_test = np.cos(X_test) + 10.0
 
 Sample functions from a prior distribution, which is zero mean.
+As shown in the figure below, `num_samples` smooth functions are sampled.
 
 .. code-block:: python
 
@@ -56,6 +66,8 @@ Sample functions from a prior distribution, which is zero mean.
     :alt: gp_sampled_prior
 
 Build a Gaussian process regression model with fixed hyperparameters.
+Fixed hyperparameters are brought through `get_hyps`.
+`mu`, `sigma`, and `Sigma` are mean estimates, standard deviation estimates, and covariance estimates, respectively.
 Then, plot the result.
 
 .. code-block:: python
@@ -108,6 +120,9 @@ Build a Gaussian process regression model with the hyperparameters optimized by 
     :alt: gp_sampled_optimized
 
 Declare some functions that would be employed as prior functions.
+`cosine` is a prior function with a cosine function.
+`linear_down` is a prior function with a decreasing function.
+`linear_up` is a prior function with an increasing function.
 
 .. code-block:: python
 
@@ -143,31 +158,6 @@ Make an another synthetic dataset using a cosine function.
 Build Gaussian process regression models with the prior functions we declare above and the hyperparameters optimized by marginal likelihood maximization, and plot the result.
 
 .. code-block:: python
-
-    def cosine(X):
-        return np.cos(X)
-
-    def linear_down(X):
-        list_up = []
-        for elem_X in X:
-            list_up.append([-0.5 * np.sum(elem_X)])
-        return np.array(list_up)
-
-    def linear_up(X):
-        list_up = []
-        for elem_X in X:
-            list_up.append([0.5 * np.sum(elem_X)])
-        return np.array(list_up)
-
-    X_train = np.array([
-        [-3.0],
-        [-2.0],
-        [-1.0],
-    ])
-    Y_train = np.cos(X_train) + 2.0
-    X_test = np.linspace(-3, 6, num_test)
-    X_test = X_test.reshape((num_test, 1))
-    Y_test = np.cos(X_test) + 2.0
 
     prior_mu = cosine
     mu, sigma, Sigma = gp.predict_with_optimized_hyps(X_train, Y_train, X_test,
