### 2023-08-20 191f35b
Update gp.rst
