### 2023-08-20 bf6a7c0
Update about_bayeso.rst and gp.rst
