+++ docs/about/about_bayeso.rst
@@ -11,7 +11,6 @@ Supported Python Version
 
 We test our package in the following versions.
 
-- Python 3.6
 - Python 3.7
 - Python 3.8
 - Python 3.9
