+++ docs/example/gp.rst
@@ -2,7 +2,12 @@ Building Gaussian Process Regression
 ====================================
 
 This example is for building Gaussian process regression models.
-First of all, import the packages we need and **bayeso**.
+In this example, we cover three scenarios:
+Gaussian process with fixed hyperparameters,
+Gaussian process with learned hyperparameters,
+and Gaussian process with particular priors.
+
+First of all, import the package we need and **bayeso**.
 
 .. code-block:: python
 
@@ -14,6 +19,9 @@ First of all, import the packages we need and **bayeso**.
     from bayeso.utils import utils_plotting
 
 Declare some parameters to control this example.
+`use_tex` is a flag for using a LaTeX style,
+`num_test` is the number of test data points,
+and `str_cov` is a kernel choice.
 
 .. code-block:: python
 
@@ -21,7 +29,8 @@ Declare some parameters to control this example.
     num_test = 200
     str_cov = 'matern52'
 
-Make a simple synthetic dataset, which produces with cosine functions.
+Make a simple synthetic dataset, which is produced with a cosine function.
+The underlying true function is \cos(x) + 10.
 
 .. code-block:: python
 
@@ -39,6 +48,7 @@ Make a simple synthetic dataset, which produces with cosine functions.
     Y_test = np.cos(X_test) + 10.0
 
 Sample functions from a prior distribution, which is zero mean.
+As shown in the figure below, `num_samples` smooth functions are sampled.
 
 .. code-block:: python
 
@@ -56,6 +66,9 @@ Sample functions from a prior distribution, which is zero mean.
     :alt: gp_sampled_prior
 
 Build a Gaussian process regression model with fixed hyperparameters.
+Fixed hyperparameters are brought through `get_hyps`.
+`mu`, `sigma`, and `Sigma` are mean estimates, standard deviation estimates, and covariance estimates, respectively.
+In addition, `num_samples` functions are sampled using `mu` and `Sigma`.
 Then, plot the result.
 
 .. code-block:: python
@@ -83,6 +96,7 @@ Then, plot the result.
     :alt: gp_sampled_fixed
 
 Build a Gaussian process regression model with the hyperparameters optimized by marginal likelihood maximization, and plot the result.
+Similar to the aforementioned case, `num_samples` functions are sampled from the distributions with `mu` and `Sigma`.
 
 .. code-block:: python
 
@@ -108,6 +122,9 @@ Build a Gaussian process regression model with the hyperparameters optimized by 
     :alt: gp_sampled_optimized
 
 Declare some functions that would be employed as prior functions.
+`cosine` is a prior function with a cosine function.
+`linear_down` is a prior function with a decreasing function.
+`linear_up` is a prior function with an increasing function.
 
 .. code-block:: python
 
@@ -127,6 +144,7 @@ Declare some functions that would be employed as prior functions.
         return np.array(list_up)
 
 Make an another synthetic dataset using a cosine function.
+The true function is \cos(x) + 2.
 
 .. code-block:: python
 
@@ -141,33 +159,9 @@ Make an another synthetic dataset using a cosine function.
     Y_test = np.cos(X_test) + 2.0
 
 Build Gaussian process regression models with the prior functions we declare above and the hyperparameters optimized by marginal likelihood maximization, and plot the result.
+Also, `num_samples` functions are sampled from the distributions defined with `mu` and `Sigma`.
 
 .. code-block:: python
-
-    def cosine(X):
-        return np.cos(X)
-
-    def linear_down(X):
-        list_up = []
-        for elem_X in X:
[…]
