### 2023-08-20 ec102dc
Add EXAMPLES.md and TESTS.md
