### 2023-08-20 80f3ef9
Update about_bayeso.rst
