### 2023-08-20 890dd6f
Update about_bayeso.rst
