### 2023-08-20 0a86413
add lists of examples and tests
