+++ examples/05_benchmarks/example_benchmarks_bohachevsky_bo_ei.py
@@ -21,51 +21,47 @@ obj_fun = Bohachevsky()
 def fun_target(X):
     return obj_fun.output(X)
 
-def main(path_save):
-    num_bo = 5
-    num_iter = 10
-    num_init = 5
+path_save = None
 
-    bounds = obj_fun.get_bounds()
+if path_save is not None and not os.path.isdir(path_save):
+    os.makedirs(path_save)
 
-    model_bo = wrappers.BayesianOptimization(
-        bounds,
-        fun_target,
-        num_iter,
-        str_surrogate='gp',
-        str_cov='matern52',
-        str_acq='ei',
-        str_initial_method_bo='sobol',
-        str_sampling_method_ao='sobol',
-        str_optimizer_method_gp='BFGS',
-        str_optimizer_method_bo='L-BFGS-B',
-        num_samples_ao=100,
-        debug=True,
-    )
+num_bo = 5
+num_iter = 10
+num_init = 5
 
-    list_Y = []
-    list_time = []
-    for ind_bo in range(0, num_bo):
-        print('BO Round', ind_bo + 1)
-        X, Y, time_all, _, _ = model_bo.optimize(num_init, seed=42 * ind_bo)
-        list_Y.append(Y)
-        list_time.append(time_all)
+bounds = obj_fun.get_bounds()
 
-        bx_best, y_best = utils_bo.get_best_acquisition_by_history(X, Y)
-        print(bx_best, y_best)
+model_bo = wrappers.BayesianOptimization(
+    bounds,
+    fun_target,
+    num_iter,
+    str_surrogate='gp',
+    str_cov='matern52',
+    str_acq='ei',
+    str_initial_method_bo='sobol',
+    str_sampling_method_ao='sobol',
+    str_optimizer_method_gp='BFGS',
+    str_optimizer_method_bo='L-BFGS-B',
+    num_samples_ao=100,
+    debug=True,
+)
 
-    Ys = np.array(list_Y)
-    Ys = np.expand_dims(np.squeeze(Ys), axis=0)
-    times = np.array(list_time)
-    times = np.expand_dims(times, axis=0)
+list_Y = []
+list_time = []
+for ind_bo in range(0, num_bo):
+    print('BO Round', ind_bo + 1)
+    X, Y, time_all, _, _ = model_bo.optimize(num_init, seed=42 * ind_bo)
+    list_Y.append(Y)
+    list_time.append(time_all)
 
-    utils_plotting.plot_minimum_vs_iter(Ys, [STR_FUN_TARGET], num_init, True, path_save=path_save, str_postfix=STR_FUN_TARGET)
-    utils_plotting.plot_minimum_vs_time(times, Ys, [STR_FUN_TARGET], num_init, True, path_save=path_save, str_postfix=STR_FUN_TARGET)
+    bx_best, y_best = utils_bo.get_best_acquisition_by_history(X, Y)
+    print(bx_best, y_best)
 
+Ys = np.array(list_Y)
+Ys = np.expand_dims(np.squeeze(Ys), axis=0)
+times = np.array(list_time)
+times = np.expand_dims(times, axis=0)
 
-if __name__ == '__main__':
-    path_save = None
-
-    if path_save is not None and not os.path.isdir(path_save):
-        os.makedirs(path_save)
-    main(path_save)
+utils_plotting.plot_minimum_vs_iter(Ys, [STR_FUN_TARGET], num_init, True, path_save=path_save, str_postfix=STR_FUN_TARGET)
+utils_plotting.plot_minimum_vs_time(times, Ys, [STR_FUN_TARGET], num_init, True, path_save=path_save, str_postfix=STR_FUN_TARGET)
