### 2023-08-20 #5320 @jungtaekkim
@salrm8 We thank you for your constructive comments!

We have revised our submission.  Please see the updated document.  However, we could not understand "Figure 1: Assuming the iterations did not stop at 9, please mention this in the caption for clarity."  This figure shows the Bayesian optimization results until 9 iterations; we can stop it here or keep going, depending on an iteration budget given.

We also updated our repository:

• Please add a list of tests: See [this link](https://github.com/jungtaekkim/bayeso/commit/0a8641344713db9efe15b5e99c26700fd4201e74).
• Add a list of examples: See [this link](https://github.com/jungtaekkim/bayeso/commit/0a8641344713db9efe15b5e99c26700fd4201e74).
• Please make the tests automatic: Our unit tests are automatically run with GitHub Actions.
• In `/exmaples/02_surrogates/`, there is an error when running `example_gp_mml_comparisons.py`:
  `ModuleNotFoundError: No module named 'bayeso_benchmarks'`: You need to install the packages described in `requirements-examples.txt`; we separated them in order to implement our software with compact dependency.  It is described in `README.md`.
• The same error in /examples/07_wandb/, when running wandb_branin.py: You need to install the packages described in `requirements-examples.txt`; we separated them in order to implement our software with compact dependency.  It is described in `README.md`.
• The same error for notebook `example_bo_branin.ipynb`: You need to install the packages described in `requirements-examples.txt`; we separated them in order to implement our software with compact dependency.  It is described in `README.md`.
• Is `xgboost` among the requirements? It is not installed automatically although it is needed for running `example_hpo_xgboost_ei.py`: You need to install the packages described in `requirements-examples.txt`; we separated them in order to implement our software with compact dependency.  It is described in `README.md`.
• Could the authors provide an example (surrogate and BO) for a 2D parameter space? It seems, the examples are only based on one parameter: Examples in `05_benchmarks` and `06_hpo` demonstrate the problems for more than one-dimensional spaces.
• Adding a bit of text in the jupyter notebooks in `/examples/99_notebooks/` would be helpful to understand the topic of each notebook: We have updated them; please see `examples/99_notebooks`.
• The documentation is good, but can be improved by adding a list for the examples and considering the previous point. Another improvement would be adding a bit of theory: We have added a list of examples and a list of tests in the documentation.
