### 2023-08-27 634d596
Update about_bayeso.rst
