+++ docs/about/about_bayeso.rst
@@ -26,46 +26,46 @@ We provide a list of examples.
 
     examples/
     ├── 01_basics
-    │   ├── example_basics_bo.py
-    │   └── example_basics_gp.py
+    │   ├── example_basics_bo.py: a basic example of Bayesian optimization
+    │   └── example_basics_gp.py: a basic example of Gaussian processes
     ├── 02_surrogates
-    │   ├── example_generic_trees.py
-    │   ├── example_gp_mml_comparisons.py
-    │   ├── example_gp_mml_kernels.py
-    │   ├── example_gp_mml_many_points.py
-    │   ├── example_gp_mml_y_scales.py
-    │   ├── example_gp_priors.py
-    │   ├── example_random_forest.py
-    │   └── example_tp_mml_kernels.py
+    │   ├── example_generic_trees.py: an example of modeling generic trees
+    │   ├── example_gp_mml_comparisons.py: an example of Gaussian processes with different optimization methods for marginal likelihood maximization
+    │   ├── example_gp_mml_kernels.py: an example of Gaussian processes with different kernels
+    │   ├── example_gp_mml_many_points.py: an example of Gaussian processes with many data points
+    │   ├── example_gp_mml_y_scales.py: an example of Gaussian processes with different scales of function evaluations
+    │   ├── example_gp_priors.py: an example of Gaussian processes with differnt prior functions
+    │   ├── example_random_forest.py: an example of modeling random forests
+    │   └── example_tp_mml_kernels.py: an example of student-$t$ processes with different kernels
     ├── 03_bo
-    │   ├── example_bo_aei.py
-    │   ├── example_bo_ei.py
-    │   ├── example_bo_pi.py
-    │   ├── example_bo_pure_exploit.py
-    │   ├── example_bo_pure_explore.py
-    │   └── example_bo_ucb.py
+    │   ├── example_bo_aei.py: an example of Bayesian optimization with augmented expected improvement
+    │   ├── example_bo_ei.py: an example of Bayesian optimization with expected improvement
+    │   ├── example_bo_pi.py: an example of Bayesian optimization with the probability of improvement
+    │   ├── example_bo_pure_exploit.py: an example of Bayesian optimization with pure exploitation
+    │   ├── example_bo_pure_explore.py: an example of Bayesian optimization with pure exploration
+    │   └── example_bo_ucb.py: an example of Bayesian optimization with Gaussain process upper confidence bound
     ├── 04_bo_with_surrogates
-    │   ├── example_bo_w_gp.py
-    │   └── example_bo_w_tp.py
+    │   ├── example_bo_w_gp.py: an example of Bayesian optimization with Gaussian process surrogates
+    │   └── example_bo_w_tp.py: an example of Bayesian optimization with student-$t$ process surrogates
     ├── 05_benchmarks
-    │   ├── example_benchmarks_ackley_bo_ei.py
-    │   ├── example_benchmarks_bohachevsky_bo_ei.py
-    │   ├── example_benchmarks_branin_bo_ei.py
-    │   ├── example_benchmarks_branin_gp.py
-    │   ├── example_benchmarks_branin_ts.py
-    │   └── example_benchmarks_hartmann6d_bo_ei.py
+    │   ├── example_benchmarks_ackley_bo_ei.py: an example of Bayesian optimization for the Ackley function
+    │   ├── example_benchmarks_bohachevsky_bo_ei.py: an example of Bayesian optimization for the Bohachevsky function
[…]
