### 2023-08-27 57731fd
Update about_bayeso.rst
