### 2023-08-27 b7bbe1a
Update about_bo.rst
