### 2023-08-27 075fb4c
Update about_bo.rst
