+++ docs/about/about_bo.rst
@@ -9,10 +9,9 @@ About Bayesian Optimization
 Bayesian optimization is a *global optimization* strategy for *black-box* and *expensive-to-evaluate* functions.
 Generic Bayesian optimization follows these steps:
 
-#. Build a *surrogate function* $\hat{f}$ with historical inputs $\mathbf{X}$ and their observations $\mathbf{y}$, which is defined with mean and variance functions.
-$$
-\hat{f}(\mathbf{x} \mid \mathbf{X}, \mathbf{y}) = \mathcal{N}(\mu(\mathbf{x} \mid \mathbf{X}, \mathbf{y}), \sigma^2(\mathbf{x} \mid \mathbf{X}, \mathbf{y}))
-$$
+#. Build a *surrogate function* :math:`\hat{f}` with historical inputs :math:`\mathbf{X}` and their observations :math:`\mathbf{y}`, which is defined with mean and variance functions.
+.. math::
+    \hat{f}(\mathbf{x} \mid \mathbf{X}, \mathbf{y}) = \mathcal{N}(\mu(\mathbf{x} \mid \mathbf{X}, \mathbf{y}), \sigma^2(\mathbf{x} \mid \mathbf{X}, \mathbf{y}))
 #. Compute and maximize an *acquisition function*, defined by the outputs of surrogate function.
 #. Observe the *maximizer* of acquisition function from a true objective function.
 #. Accumulate the maximizer and its observation.
