### 2023-08-27 7a65f3c
Update about_bo.rst
