+++ docs/about/about_bo.rst
@@ -9,13 +9,14 @@ About Bayesian Optimization
 Bayesian optimization is a *global optimization* strategy for *black-box* and *expensive-to-evaluate* functions.
 Generic Bayesian optimization follows these steps:
 
-#. Build a *surrogate function* :math:`\hat{f}` with historical inputs :math:`\mathbf{X}` and their observations :math:`\mathbf{y}`, which is defined with mean and variance functions.
+1. Build a *surrogate function* :math:`\hat{f}` with historical inputs :math:`\mathbf{X}` and their observations :math:`\mathbf{y}`, which is defined with mean and variance functions.
+
 .. math::
 
     \hat{f}(\mathbf{x} \mid \mathbf{X}, \mathbf{y}) = \mathcal{N}(\mu(\mathbf{x} \mid \mathbf{X}, \mathbf{y}), \sigma^2(\mathbf{x} \mid \mathbf{X}, \mathbf{y}))
-#. Compute and maximize an *acquisition function*, defined by the outputs of surrogate function.
-#. Observe the *maximizer* of acquisition function from a true objective function.
-#. Accumulate the maximizer and its observation.
+2. Compute and maximize an *acquisition function*, defined by the outputs of surrogate function.
+3. Observe the *maximizer* of acquisition function from a true objective function.
+4. Accumulate the maximizer and its observation.
 
 This project helps us to execute this Bayesian optimization procedure.
 In particular, *Gaussian process regression* is used as a surrogate function,
