### 2023-08-27 47f54c6
Update about_bo.rst
