### 2023-08-27 accb2cf
Update about_bo.rst
