### 2023-08-28 79b97b8
Update about_bo.rst
