### 2023-08-28 dbdd11d
Update about_bo.rst
