### 2023-08-28 #5320 @jungtaekkim
Thank you @salrm8!  We have updated our documentation based on your comment.

- Added description on examples and tests; you can see [this link](https://bayeso.readthedocs.io/en/0.5.4/about/about_bayeso.html) and https://github.com/jungtaekkim/bayeso/commit/57731fd28b87a90fc16ba98dfe0f2ed215d881ed.
- Added some mathematical expressions; you can see [this link](https://bayeso.readthedocs.io/en/0.5.4/about/about_bo.html) and https://github.com/jungtaekkim/bayeso/commit/dbdd11da5d5bc87b158d42ea51360415aac06dd8.
