### 2023-08-28 36436c3
Update about_bo.rst
