### 2023-08-28 1931a4f
Update about_bayeso.rst and about_bo.rst
