### 2023-08-28 f72a06b
Update about_bo.rst
