### 2023-09-09 9708ba1
Update paper.md; fix capitalization of section titles
