### 2023-10-08 19b4019
Update and fix the bib file
