+++ paper.bib
@@ -181,15 +181,6 @@
     doi={10.1016/j.artint.2013.10.003}
 }
 
-@article{MockusJ1978tgo,
-    author={{Mo\v{c}kus}, J. and {Tiesis}, V. and {\v{Z}ilinskas}, A.},
-    title={The Application of {Bayesian} Methods for Seeking the Extremum},
-    journal={Towards Global Optimization},
-    volume=2,
-    pages={117--129},
-    year=1978
-}
-
 @article{JonesDR1998jgo,
     author={{Jones}, D. R. and {Schonlau}, M. and {Welch}, W. J.},
     title={Efficient Global Optimization of Expensive Black-Box Functions},
@@ -212,7 +203,7 @@
 }
 
 @article{HuangD2006jgo,
-    author={{Huang}, D. and {Allen}, T. T. and {Notz}, W. I. and {Zeng}, N.},
+    author={{Huang}, D. and {Allen}, T. T. and {Notz}, W. I. and {Zheng}, N.},
     title={Global optimization of stochastic black-box systems via sequential kriging meta-models},
     journal={Journal of Global Optimization},
     volume=34,
@@ -334,5 +325,16 @@
     booktitle={Advances in Neural Information Processing Systems (NeurIPS)},
     volume=33,
     pages={21524--21538},
-    year=2020,
+    year=2020
+}
+
+@article{KushnerHJ1964jbe,
+    author={Kushner, H. J.},
+    title={A New Method of Locating the Maximum Point of an Arbitrary Multipeak Curve in the Presence of Noise},
+    journal={Journal of Basic Engineering},
+    volume=86,
+    number=1,
+    pages={97--106},
+    year=1964,
+    doi={10.1115/1.3653121}
 }
