+++ .readthedocs.yml
@@ -9,7 +9,7 @@ formats:
   - pdf
 
 python:
-  version: 3.11
+  version: 3.8
   install:
     - requirements: requirements.txt
     - requirements: requirements-optional.txt
