### 2023-10-08 #5320 @jungtaekkim
A tagged release is v0.5.4 ([release link](https://github.com/jungtaekkim/bayeso/releases/tag/v0.5.4)),

and DOI is 10.5281/zenodo.8419023 ([DOI link](http://doi.org/10.5281/zenodo.8419023)).
