### 2023-10-08 #5320 @jungtaekkim
I think that @salrm8 did not check a `Human and animal research` box in the review.

@drvinceknight please look at it if check boxes should be all checked.
